+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# SPMediaKeyTap CHANGELOG
+
+## 0.1.0
+
+Initial release.
