+++ Classes/osx/SPInvocationGrabbing/NSObject+SPInvocationGrabbing.h
