+++ Classes/osx/SPInvocationGrabbing/NSObject+SPInvocationGrabbing.m
