+++ Classes/osx/SPMediaKeyTap.h
