+++ Project/SPMediaKeyTapExample.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,1 @@
+<?xml version='1.0' encoding='UTF-8'?><Workspace version='1.0'><FileRef location='group:SPMediaKeyTapExample.xcodeproj'/><FileRef location='group:Pods/Pods.xcodeproj'/></Workspace>
